fix(catalog): fix(catalog): accumulate SELECT-view SUM through sumAccumulator for exact integer results
computeAggregateValue and computeViewAggregate accumulated SUM in
float64 via toFloat64, losing integer precision above 2^53 on the
SELECT-view and materialized-view paths while the GROUP BY path was
exact. Both now accumulate through sumAccumulator (DISTINCT
materialization and nil/error gating preserved). The coverage-perf
test's SUM pin updates to the corrected int64-while-integral result
type, mirroring the file's own int64 COUNT convention (hunt round 10).

+++ pkg/catalog/coverage_perf_test.go
@@ -190,7 +190,7 @@ func TestDeletedRowSkippingCoverage(t *testing.T) {
 	}
 
 	res = covQuery(t, c, "SELECT SUM(val) FROM cov_del")
-	if res.Rows[0][0] != float64(40) {
+	if res.Rows[0][0] != int64(40) {
 		t.Errorf("SUM after delete: got %v, want 40", res.Rows[0][0])
 	}
 }
